Fix running finalize script.
According to the [man page](https://github.com/systemd/mkosi/blob/master/mkosi.md#files),
the mkosi.finalize script is invoked as last step of preparing an image, from
the host system. This commit fixes the location of the `run_finalize_script()`
function call to run after the image is created. We also add in 2 more
environment variables that get passed to the mkosi.finalize script: OUTPUT_DIR
and OUTPUT.